Move a general hook function to a separate file
